+++ gemrb/plugins/MUSImporter/MUSImp.h
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/MUSImporter/MUSImp.h,v 1.12 2004/02/24 22:20:40 balrog994 Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/MUSImporter/MUSImp.h,v 1.13 2004/08/10 19:11:30 guidoj Exp $
  *
  */
 
@@ -35,7 +35,7 @@ typedef struct PLString {
 	char PLLoop[10];
 	char PLTag[10];
 	char PLEnd[10];
-	unsigned long soundID;
+	unsigned int soundID;
 } PLString;
 
 class MUSImp : public MusicMgr {
@@ -46,7 +46,7 @@ private:
 	int PLpos, PLnext;
 	FileStream* str;
 	std::vector< PLString> playlist;
-	unsigned long lastSound;
+	unsigned int lastSound;
 private:
 	void PlayMusic(int pos);
 	void PlayMusic(char* name);
